backend/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
psycopg2-binary
python-dotenv
passlib[bcrypt]
pydantic[email]
statsmodels
python-multipart
scikit-learn
python-jose[cryptography]
passlib[bcrypt]
